// web/apps/webapp/src/queries/__generated__/organizationUser_updateMyBillingContactDetailsMutation.graphql.ts
export type UpdateMyBillingContactDetailsInput = {
  addressLine1?: string | null | undefined;
  addressLine2?: string | null | undefined;
  city?: string | null | undefined;
  clientMutationId?: string | null | undefined;
  companyName?: string | null | undefined;
  country?: string | null | undefined;
  email?: string | null | undefined;
  province?: string | null | undefined;
  suburb?: string | null | undefined;
  zipcode?: string | null | undefined;
};
export type organizationUser_updateMyBillingContactDetailsMutation$variables = {
  input: UpdateMyBillingContactDetailsInput;
};
export type organizationUser_updateMyBillingContactDetailsMutation$data = {
  readonly updateMyBillingContactDetails: {
    readonly customerBillingContactDetails: {
      readonly addressLine1: string | null | undefined;
      readonly addressLine2: string | null | undefined;
      readonly city: string | null | undefined;
      readonly companyName: string | null | undefined;
      readonly country: string | null | undefined;
      readonly email: string | null | undefined;
      readonly id: string;
      readonly province: string | null | undefined;
      readonly suburb: string | null | undefined;
      readonly zipcode: string | null | undefined;
    };
  } | null | undefined;
};
export type organizationUser_updateMyBillingContactDetailsMutation$rawResponse = {
  readonly updateMyBillingContactDetails: {
    readonly customerBillingContactDetails: {
      readonly addressLine1: string | null | undefined;
      readonly addressLine2: string | null | undefined;
      readonly city: string | null | undefined;
      readonly companyName: string | null | undefined;
      readonly country: string | null | undefined;
      readonly email: string | null | undefined;
      readonly id: string;
      readonly province: string | null | undefined;
      readonly suburb: string | null | undefined;
      readonly zipcode: string | null | undefined;
    };
  } | null | undefined;
};
export type organizationUser_updateMyBillingContactDetailsMutation = {
  rawResponse: organizationUser_updateMyBillingContactDetailsMutation$rawResponse;
  response: organizationUser_updateMyBillingContactDetailsMutation$data;
  variables: organizationUser_updateMyBillingContactDetailsMutation$variables;
};